requirements.txt
requests>=2.19.1
google-cloud-storage>=1.12.0
google-cloud-logging>=1.7.0